# Conflicts:
#	src/UI/Headquarters/WB.UI.Headquarters/HqApp/src/hqapp/Views/HQ/index.js
#	src/UI/Headquarters/WB.UI.Headquarters/WB.UI.Headquarters.csproj

+++ src/UI/Headquarters/WB.UI.Headquarters/API/WebInterview/IStatefullInterviewSearcher.cs
@@ -4,7 +4,7 @@ namespace WB.UI.Headquarters.API.WebInterview
 {
     public interface IStatefullInterviewSearcher
     {
-        SearchResults Search(IStatefulInterview interview, FilterOption[] flags, long skip, long take);
+        SearchResults Search(IStatefulInterview interview, FilterOption[] flags, int skip, int take);
     }
 
     public enum FilterOption
